% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.auralLandsape.R
\name{plot.auralLandscape}
\alias{plot.auralLandscape}
\title{Plot an aural landscape from a structured dataset}
\usage{
\method{plot}{auralLandscape}(
  x,
  plotType = c("hull", "points"),
  dateRange = c(1:12),
  timeRange = c(3:22),
  observers = "all",
  obsGrep = FALSE,
  taxa = "all",
  taxaGrep = FALSE,
  show = TRUE
)
}
\arguments{
\item{x}{\code{data.frame} of observations, with taxon, day, month, year, time (in 24 hr notation)}

\item{dateRange}{Vector of months to include}

\item{timeRange}{Vector of hours to include (24 hr notation)}

\item{observers}{Character vector of length 1; observers to include, or 'all'}

\item{obsGrep}{Boolean; use \code{grep} with \code{observers}?}

\item{taxa}{Character vector of length 1; taxa to include, or 'all'}

\item{taxaGrep}{Boolean; use \code{grep} with \code{taxa}?}

\item{showPlot}{Boolean; plot or just return object?}
}
\description{
\code{plot.auralLandscape} currently does a simple 2-d plot
of observations by time of day and day of year. Eventually
should have functions to (1) accommodate point observations
and ranges, (2) subset by observers and areas, (3) allow
for plotting customization.
}
