% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAuralLandscape.R
\name{plotAuralLandscape}
\alias{plotAuralLandscape}
\title{Plot an aural landscape from a structured dataset}
\usage{
plotAuralLandscape(
  x,
  plotType = c("hull", "points"),
  addJitter = FALSE,
  ptSize = 4,
  nullSize = 2,
  legPos = c(0.85, 0.85),
  dateRange = c(1:12),
  timeRange = c(3:22),
  observers = "all",
  obsGrep = TRUE,
  taxa = "all",
  taxaGrep = TRUE,
  addAll = "gray85",
  aggregateYears = TRUE,
  timeAsDecimal = TRUE,
  colorAlpha = 0.6,
  taxonPalette = c("#000000", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7"),
  showPlot = TRUE
)
}
\arguments{
\item{x}{\code{data.frame} of observations, with taxon, day, month, year, time (in 24 hr notation)}

\item{plotType}{not currently implemented}

\item{addJitter}{Boolean; jitter points for additional visibility?}

\item{ptSize}{Point size for selected observations}

\item{nullSize}{Point size for all observations, if \code{addAll = T}}

\item{legPos}{Legend position; if \code{NA}, default}

\item{dateRange}{Vector of months to include}

\item{timeRange}{Vector of hours to include (24 hr notation)}

\item{observers}{Character vector of length 1; observers to include, or 'all'}

\item{obsGrep}{Boolean; use \code{grep} with \code{observers}?}

\item{taxa}{Character vector of length 1; taxa to include, or 'all'}

\item{taxaGrep}{Boolean; use \code{grep} with \code{taxa}?}

\item{addAll}{\code{NA} or named color; if color, add all obs in background}

\item{aggregateYears}{Ignore year when plotting, aggregate all dates to yr = 1}

\item{timeAsDecimal}{Boolean; time is decimal? alternative: H:M (24 hour)}

\item{colorAlpha}{Transparency of points}

\item{taxonPalette}{Color palette for taxa; default is color-blind with black}

\item{showPlot}{Boolean; plot or just return object?}
}
\description{
\code{plotAuralLandscape} currently does a simple 2-d plot
of observations by time of day and day of year. Eventually
should have functions to (1) accommodate point observations
and ranges, (2) subset by observers and areas, (3) allow
for plotting customization.
}
\examples{
## data(auralObservations)
## get most recent data:
require(openxlsx)
a <- read.xlsx('https://github.com/andrew-hipp/auralLandscapes/blob/main/data/auralObservations.xlsx?raw=true',
            detectDates=T)
## use this if you are using decimal time:
a$time <- a$time * 24
## use this if you want to convert decimal time to HM:
\dontrun{
  a$time <- format(ISOdatetime(1900,1,1,0,0,0, tz="GMT") +
           as.difftime(a$time, unit="hours"), "\%H:\%M")
}
out <- plotAuralLandscape(a, taxa = 'robin|cardinal|peeper|chorus')

}
